cleaning up cmes models
