+++ src/twpasolver/modes_rwa.py
@@ -1142,13 +1142,6 @@ class ModeArrayFactory:
             modes.append(Mode(label=f"i{n}", direction=direction))
             relations.append([f"i{n}", "i+" * (n - 1) + "i"])
 
-        # for n in range(2, n_sidebands + 2):
-        #     modes.append(Mode(label=f"s{n}p", direction=direction))
-        #     modes.append(Mode(label=f"i{n}p", direction=direction))
-        #     relations.append([f"s{n}p", "s+s-p"])
-        #     relations.append([f"i{n}p", "i+i-p"])
-       # modes.append(Mode(label=f"d", direction=direction))
-       # relations.append(["d", "s-p"])
         return ModeArray(modes, relations, interpolator)
 
     @staticmethod
